+++ backend/src/routes/report.js
@@ -6,25 +6,25 @@ const routes = [
   {
     method: "GET",
     url: `/${what}/projects/:projectId`,
-    schema: validators.getOneValidator,
+    schema: validators.getOne,
     handler: controller.getOne,
   },
   {
     method: "GET",
     url: `/${what}/projects/ProjectBudgetReport`,
-    schema: validators.getOneValidator,
+    schema: validators.getOne,
     handler: controller.getProjectBudgetReport,
   },
   {
     method: "GET",
     url: `/${what}/projects/ProjectQuarterlyReport`,
-    schema: validators.getOneValidator,
+    schema: validators.getOne,
     handler: controller.getProjectQuarterlyReport,
   },
   {
     method: "GET",
     url: `/${what}/projects/ProjectStatusReport`,
-    schema: validators.getOneValidator,
+    schema: validators.getOne,
     handler: controller.getProjectStatusReport,
   },
 ];