+++ backend/src/validators/report.js
@@ -1,21 +1,11 @@
-const getOneValidator = {
-  // Request parameters.
-  params: {
-    projectId: { type: "string" },
-  },
-  // Response validation.
-  response: {
-    200: {
-      type: "object",
-      properties: {
-        data: {
-          projectId: { type: "integer" },
-        },
-      },
-    },
-  },
+const { Schema, getResponse } = require("./common_schema.js");
+const S = require("fluent-json-schema");
+
+const getOne = {
+  params: S.object().prop("projectId", Schema.Id),
+  response: getResponse(S.object().prop("projectId", Schema.Id)),
 };
 
 module.exports = {
-  getOneValidator,
+  getOne,
 };