+++ backend/src/validators/subcontractors.js
@@ -1,21 +1,31 @@
-const getOneValidator = {
-  // Request parameters.
-  params: {
-    id: { type: "string" },
-  },
-  // Response validation.
-  response: {
-    200: {
-      type: "object",
-      properties: {
-        data: {
-          id: { type: "integer" },
-        },
-      },
-    },
-  },
+const { Schema, getResponse, getAddResponse, getUpdateResponse } = require("./common_schema.js");
+const S = require("fluent-json-schema");
+
+const body = S.object().prop("id", Schema.Id).prop("subcontractor_name", Schema.ShortString);
+
+const getAll = {
+  response: getResponse(S.array().items(body)),
+};
+
+const getOne = {
+  params: Schema.IdParam,
+  response: getResponse(body),
+};
+
+const updateOne = {
+  params: Schema.IdParam,
+  body: body.without(["id"]).minProperties(1),
+  response: getUpdateResponse(),
+};
+
+const addOne = {
+  body: body.without(["id"]).required(["subcontractor_name"]),
+  response: getAddResponse(),
 };
 
 module.exports = {
-  getOneValidator,
+  getAll,
+  getOne,
+  updateOne,
+  addOne,
 };